requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
python-dotenv==1.0.1
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
numpy==1.26.4
requests==2.32.3
Pillow==10.4.0
# On Raspberry Pi, install manually:
# pip install https://github.com/google-coral/pycoral/releases/download/release-frogfish/tflite_runtime-2.15.0-<your_python>-<your_arch>.whl
google-generativeai==0.7.2
tensorflow==2.15.0
protobuf==4.25.1
